[US_ND] YAML for new raw data (Recidiviz/recidiviz-data#27271)
## Description of the change

This adds a raw data config for a table called
`elite_offender_medical_screenings_6i` that has been filtered of all
medically-sensitive information. This is for some reason where ND stores
information about appearances in front of the parole board. There is a
specific "medical questionnaire code" (`PAR`) that denotes when the row
is actually about a parole board appearance and not a medical
evaluation. The table is filtered to include only those rows before we
receive it.

I loaded the data to a sandbox to test that the config works as
expected.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7247caa30eb77f9fd0f61ae9933d4d4080959265