[BigQuery] Update external validation views to use UnionAllBigQueryViewBuilder (Recidiviz/recidiviz-data#27266)
## Description of the change

This PR updates the `UnionAllBigQueryViewBuilder` to omit parent views
during the build() step that are not deployed to the current project.
This allows us to migrate our external validation views to use
`UnionAllBigQueryViewBuilder` and avoid relying on a call to
`metadata.project_id()` at view builder collection time.

I ran the `deploy_test_empty_views` script from this branch against both
prod and staging and it succeeded.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

See this thread:
https://recidiviz.slack.com/archives/C028X32LRH7/p1706815401113279

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 49a8f4b295c2061406be5a0b80141ade9629ebee