[US_ND] Integrate bed assignments into IPs (Recidiviz/recidiviz-data#24748)
## Description of the change

This uses `elite_bedassignmenthistory` and `elite_livingunits` to add a
level of detail to incarceration periods in ND. Since the view is
updated to include internal movements, the old view named
`elite_externalmovements_incarceration_periods` is deprecated and
replaced with a new view with all the same functionality called
`elite_movements_incarceration_periods`.

For an example of the new structure, let's say a person was admitted to
county jail on 1/1/2018, then sentenced to 3 months in NDSP on
1/15/2018. In `elite_bedassignmenthistory`, we would find records of
them in a holding cell in county jail for 15 days between their arrest
and sentencing, then in an administrative holding cell at NDSP for 2
days while a bed was found for them, and then the rest of their sentence
in their permanent bed assignment at NDSP. The new/relevant fields in
their incarceration period in the state dataset would look like
[this](https://docs.google.com/spreadsheets/d/1zuDJSl1rcIQjns0Fw1-Ahd8pS7ZxWmcbqt5_W0nkBoM/edit?usp=sharing).

One placeholder in these changes:
- We are making assumptions as to which `elite_livingunits` assignments
mean people are in solitary confinement, and which types of solitary
confinement they are (i.e. disciplinary, administrative, etc). We know
some of them, but there are a lot of codes. We have asked the BIU team
in ND for help categorizing, and in the meantime have made a ticket to
track this work here:
https://github.com/Recidiviz/recidiviz-data/issues/26928.

Tests:

- [x] successful ingest pipeline run
([link](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-01-22_11_31_33-2134078370588550478;graphView=0?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22))))
- [x] ingest_mappings script passes
- [x] confirmed view query is deterministic

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#25026

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1ca5487179a3570f14a39ba592cb05b588c31eb3