[US_CA] Ingest PVDTS referrals as programs (Recidiviz/recidiviz-data#27004)
## Description of the change

We receive a subset of treatment referral data from the PVDTS database
maintained by DAPO that reflect only referrals to non-state sponsored
programs that actually resulted in a program start. This PR ingests
these referrals as StateProgramAssignment objects.

Please review the
[us_ca/schema_mappings.md](https://github.com/Recidiviz/recidiviz-data/pull/27004/files#diff-fe0d46e548df8474d2a0b728b4b4ea9a386deb9b3abd776c934d23ed23a1b7b9)
doc for more information on the difference between these too files (and
just to review it because you are thorough at PR reviews).

Also includes a new version of BQ emulator.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 9b3d3f3ebb79ee891c98aa6df6315fe22f4293ea