give auth0 service account the ability to call the admin panel service (Recidiviz/recidiviz-data#27312)
## Description of the change

We'd like to call the admin panel service from auth0 instead of the
recidiviz-data appengine service, since requests to the latter sometimes
time out, which we believe is due to the service being busy doing
ingest. The admin panel service has the same auth endpoints, but should
be more responsive because it doesn't handle long-running requests in
the same way.

I don't think we have a published standard for when we do IAM stuff in
TF vs access control tickets vs JIT oncall access, so I opted to put it
in TF since I think it's nice to have this documented in the code :)

Relevant plan output:
```
  # google_cloud_run_service_iam_member.admin-panel-auth0-actions will be created
  + resource "google_cloud_run_service_iam_member" "admin-panel-auth0-actions" {
      + etag     = (known after apply)
      + id       = (known after apply)
      + location = "us-central1"
      + member   = "serviceAccount:auth0-actions-service-account@recidiviz-staging.iam.gserviceaccount.com"
      + project  = "recidiviz-staging"
      + role     = "roles/run.invoker"
      + service  = "admin-panel"
    }
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

https://github.com/Recidiviz/recidiviz-dashboards/issues/4567

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: fd20a406271cfcbc8d7014ca18509c8452d03c16