[Justice Counts] skip child agency names when they are empty strings (Recidiviz/recidiviz-data#27299)
## Description of the change

[This](https://recidiviz-inc.sentry.io/issues/4932751241/?project=4504532096516096&query=is%3Aunresolved&referrer=issue-stream&stream_index=0)
key error was surfaced on Wednesday when NCDPS attempted to upload a
sheet with agency names for child agencies that are not provisioned.
When we run into a child agency name in `SpreadsheetUploader` that does
not exist, we save the name as "" and add an error to the
`metric_key_to_errors` dictionary. Later on, these names were not
filtered out as illegitimate agencies and so a key error was raised when
we attempt to index into `child_agency_name_to_agency` with "" as an
agency name. I added a small change to filter out the "" child agency
names so that we can ingest the data in the file.

How I tested: Deployed a playtesting branch in production and
re-uploaded the sheets successfully for NCDPS.

[Screen recording of video will be here once I have better wifi, but
it's a video of me uploading and getting to the errors/warnings page...]

## Related issues

Closes Recidiviz/recidiviz-data#27255

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5988e60d35013d53cb8364178a4616ebb82f6a82