[Justice Counts] Cache the return value of _SharedStateCode._state_info_map(). (Recidiviz/recidiviz-data#27263)
## Description of the change

Note: This PR is the output of a pair programming session with @morden35
and @mxosman!

Cache the return value of `_SharedStateCode._state_info_map()` so that
we are not rebuilding the Dictionary with each subsequent call. Caching
the return value here is safe because
`_SharedStateCode._state_info_map()` will always output identical maps.

In the flame chart below, we see the number of times that
`_state_info_map` is called for a single call of
`admin.fetch_all_users`. The chart looks similar for
`admin.fetch_all_agencies`.

<img width="1728" alt="Screenshot 2024-02-01 at 5 03 26 PM"
src="https://github.com/Recidiviz/recidiviz-data/assets/130382407/bae02344-cc00-4b5a-83c5-da90e8184ee7">

Some of these calls take as a long as half a second, the full
`admin.fetch_all_users` call taking 2.5 seconds.

The Loom video below shows that we can use the python caching decorator
such that we only build the full state dictionary once, and use the
cached value for subsequent calls to `_state_info_map`.

https://github.com/Recidiviz/recidiviz-data/assets/130382407/e7b7ddb2-9e45-4a3a-838d-447ef1718958

We deployed this PR to playtesting and looked through the resulting
Sentry flame chart. The screenshot below suggests that `fetch_all_users`
is now taking 0.4 ms instead of 2.5 seconds.

<img width="1148" alt="301661773-d22ff150-fbdb-4c5d-976d-24aab06b415e-1"
src="https://github.com/Recidiviz/recidiviz-data/assets/130382407/79cbe6d5-4096-43c7-9e2b-9d3c5a6c8268">

@morden35 @mxosman and I think this is an improvement over the
status-quo and that we can submit this change. But we would like to
check in again after a week or so once there is more Sentry data and
determine whether caching this method is enough of an improvement.

If we find it's not the fix we're looking for, we might consider other
options, such as building a static map of state codes to names.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#26952

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 893bf8382f9a547fb95d687bb625eeb75715003c