[Cleanup] Update latest view builder collection to not depend on project (Recidiviz/recidiviz-data#27275)
## Description of the change

Makes two minor updates to how we generate our raw data `latest` views:
1) Updates view builder collection to generate view builders for all
states that exist in the codebase, regardless of whether infrastructure
exists for that state in the current project (e.g. includes a US_OZ
builder even when building in prod), but then updates the
`projects_to_deploy` logic in the builder to select only staging for
playground states.
2) Adds a post-processing step to remove hydrated project ids from the
view query templates so that we can enforce that view query templates
are the same regardless of project (these are project_ids are hydrated
in the template at build time because we reuse a query builder helper
that is used outside of the view deploy).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

See this thread:
https://recidiviz.slack.com/archives/C028X32LRH7/p1706815401113279.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 27b5ecf35461ff9f862117202afd7a444d868858