Speed up `validate_source_visibility` check (Recidiviz/recidiviz-data#27045)
## Description of the change

This updates our `validate_source_visibility` check to rely on parsing
source files and walking the AST to find imports and build a graph of
dependencies.

The old version of this logic relied on the `modulefinder` code from the
python base library, that we vendored and altered to work for our use
case. This was really slow and brittle. Chinedu used this AST parsing
method for automatically generating the list of files used by each
Airflow DAG so that we could copy them to GCS, and that was a lot
faster. This PR generalizes that code to be used in both places.

- [ ] Run terraform plan to ensure there aren't any unexpected changes
to the files that will go to airflow

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#23809

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Helper Bot <helperbot@recidiviz.org>
Co-authored-by: Anna Geiduschek <anna@recidiviz.org>
GitOrigin-RevId: 7fc7a2477397546cc5bb4eb0da0b3e2cabfa2802