[Justice Counts] un-gate reminder emails to non-recidiviz emails (Recidiviz/recidiviz-data#27412)
## Description of the change

Caught this errant to-do while working on implementing reminder emails
for super agencies! This needs to be un-gated before so that the job
will successfully send the emails on the 15th.

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: eef75b942e5cf6bf4b6e5ef078e731cb65f436d8