[Testing] Enforce that view builder collection does not rely on project_id (Recidiviz/recidiviz-data#27279)
## Description of the change

Updates / adds tests in `deployed_view_tests.py` that enforce the
following invariants:
1) The set of view *builders* returned by
`_all_deployed_view_builders()` should not be different when run in prod
vs staging.
2) The only mechanism used to filter out views that should not deploy in
a given project should be calling `should_deploy_in_project()` on a view
builder.
3) With the exception of `UnionAllBigQueryViewBuilder` (which may filter
views out of its query based on whether they should be deployed), the
view_query_template for each built view should be identical whether it
is built in staging or prod.

Also updates these tests so that `project_id()` mocking is only done
strictly where it is needed (i.e. at view *build* step).

As part of this change, I updated the `deploy_view_builders()` function
to non longer take in the project_id - in all places outside of tests
the project_id was already set via a local override or other, so this
was redundant.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

See this thread:
https://recidiviz.slack.com/archives/C028X32LRH7/p1706815401113279

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8330a30f0a42f3ada653b6e76bd97a0beb4b9eb9