[Cleanup] Some assorted schema_utils-related cleanup (Recidiviz/recidiviz-data#27403)
## Description of the change

For the various helpers in `schema_utils.py` we have been somewhat
undisciplined about what value we pass in to determine which schema to
inspect. We sometimes pass in `SchemaType`, sometimes the schema module,
and sometimes the `DeclarativeMeta` (e.g. `StateBase`). I think it would
be great to have all functions in this file just take in a `SchemaType`
(the lightest weight / least error-prone of all of those), then derive
what we need from there. This PR brings us part of the way towards that
vision and also eliminates some of the schema-specific functions where a
call to a generic function with a `SchemaType` arg will do.

This was motivated by wanting a cleaner way to get the list of tables in
the state schema so I can simplify the logic to generate the `state`
schema from `us_xx_state_primary` datasets without having to get the
list of tables from the `CloudSqlToBQConfig`.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Prework for Recidiviz/recidiviz-data#20930.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 9b086e1b4eca43c9fd00e1b7e2ddf82747227072