[Justice Counts] do not default to uploading for superagency when child agency name is not recognized (Recidiviz/recidiviz-data#27464)
A bug was surfaced by NCDPS today when they tried to upload data for
several child agencies and they noticed data was written for their
superagency. Turns out, when we encounter an empty agency column or an
agency column with an invalid / misspelled agency name we set the agency
name as "". Later on in Bulk Upload, we don't skip these rows (even
though they should not be ingested because we don't know what agency
they are for) and in the following piece of code

https://github.com/Recidiviz/recidiviz-data/blob/ad2b73458ae880a8de8c3aec1711c2f68a0848f5/recidiviz/justice_counts/bulk_upload/spreadsheet_uploader.py#L401-L405

the unknown agency is set to the current agency uploading because the
agency name was set to "".

This PR includes two changes:

1) Skipping rows that have empty / misspelled agency names during
ingest.
2) Generating templates for superagency metrics without an agency
column, superagency metrics are only for the superagency so no child
agency should be specified for the rows.

The video below shows me uploading a sheet with an invalid agency name.
The rows associated with the invalid agency name are not ingested. It
also shows me generating a superagency template and showing that it has
no agency column. I then upload a superagency sheet and show the data
was correctly ingested.

https://github.com/Recidiviz/recidiviz-data/assets/19961693/69954b31-da1c-4739-a186-53137ccd24da

## Related issues

Closes Recidiviz/recidiviz-data#27463

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b79afc58d48c2631d68815e0cf4f7c760c73f457